requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
google-cloud-storage
google-cloud-bigquery
pandas
gcsfs
beautifulsoup4
requests
pyspark==3.3.2

# apache-airflow
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-apache-spark==4.11.3

astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy